refactor: move print styles to a dedicated components directory for better organization
