+++ src/renderer/views/Reports/components/index.ts
@@ -2,3 +2,4 @@ export * from './EmptyState';
 export * from './LoadingState';
 export * from './SortableHeader';
 export * from './useSorting';
+export * from './printStyles';
