+++ src/renderer/views/Reports/components/printStyles.ts
